Implement pause.
We now start out immigrating colleagues paused. They can still receive
direct messages from the leader that is naturalizing them. The leader
can adjust the pause, so that the leader can know exactly when the pause
begins and build the state of the colleague from a snapshot of the state
at that point. The paused messages are replayed when the immigrating
colleague is marked as naturalized.