Reduced now returns an array.
It will be easier to do what you'll most want to do, sum or count or
similar using `Array.reduce`, transform using `Array.map`, etc. The
`participantId` is not likely to be meaningful, but if I find I need it
frequently, I can tack it to the end of the callback.