Rework Conference for the new Colleague.
Removed a lot of dead code. Reduce complexity of the offering. No more
notion of using the atomic log to send messages to specific
participants. Such a think would actually be build on top of broadcast,
each participant receiving the broadcast, but only certain ones
responding, others responding that they will not be responding. There is
no need for this sort of addressed at the moment, so we'll wait until we
have use case.

Otherwise, methods are now mostly error-first callback functions. Using
the responder pattern instead of `EventEmitter`s.